+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.4.8 - Unreleased
 
+### Fixed
+* Fixed potential crash in autompl on some systems with conflicting PyQt5 and
+  python-opencv libraries. This involves falling back to agg if there *might*
+  be a conflict. The environ KWPLOT_UNSAFE=1 can be set to use the old behavior
+
 
 ## Version 0.4.7 - Released 2020-08-26